% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iivSearch.R
\name{addAllEtas}
\alias{addAllEtas}
\title{Add Individual Random Effects and Fix them to Small Value}
\usage{
addAllEtas(ui, fix = FALSE)
}
\arguments{
\item{ui}{model with no individual random effects added}

\item{fix}{If TRUE, the added etas will be fixed (not estimatable). Default is FALSE.}
}
\value{
nlmixr2 fit with individual random effects added on all fixed effects and fixed.
}
\description{
Add Individual Random Effects and Fix them to Small Value
}
\author{
Omar I. Elashkar
}
