% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsingutil.R
\name{addCatCovariates}
\alias{addCatCovariates}
\title{Make dummy variable cols and updated covarsVec}
\usage{
addCatCovariates(data, covarsVec, catcovarsVec)
}
\arguments{
\item{data}{data frame used in the analysis}

\item{covarsVec}{character vector of covariates that need to be
added}

\item{catcovarsVec}{character vector of categorical covariates that
need to be added}
}
\value{
return updated Data along with the updated covarsVec
}
\description{
Make dummy variable cols and updated covarsVec
}
\author{
Vishal Sarsani
}
