% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearizefocei.R
\name{addData2Rx}
\alias{addData2Rx}
\title{Add Data to RxUi model}
\usage{
addData2Rx(ui, data)
}
\arguments{
\item{ui}{a model.}

\item{data}{a data frame to attach.
The data can then be accessed via \code{nlme::getData(ui)}}
}
\value{
rxUi.
}
\description{
Add Data to RxUi model
}
\author{
Omar I. Elashkar
}
