% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsingutil.R
\name{addorremoveCovariate}
\alias{addorremoveCovariate}
\title{Add covariate}
\usage{
addorremoveCovariate(ui, varName, covariate, add = TRUE)
}
\arguments{
\item{ui}{compiled rxode2 nlmir2 model or fit}

\item{varName}{the variable name to which the given covariate is to be added}

\item{covariate}{the covariate that needs string to be constructed}

\item{add}{boolean indicating if the covariate needs to be added or removed.}
}
\description{
Add covariate
}
\author{
Matthew Fidler, Vishal Sarsani
}
