% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computingutil.R
\name{bootplot}
\alias{bootplot}
\alias{bootplot.nlmixr2FitCore}
\title{Produce delta objective function for boostrap}
\usage{
bootplot(x, ...)

\method{bootplot}{nlmixr2FitCore}(x, ...)
}
\arguments{
\item{x}{fit object}

\item{...}{other parameters}
}
\value{
Fit traceplot or nothing.
}
\description{
Produce delta objective function for boostrap
}
\references{
R Niebecker,  MO Karlsson. (2013)
\emph{Are datasets for NLME models large enough for a bootstrap to provide reliable parameter uncertainty distributions?}
PAGE 2013.
\url{https://www.page-meeting.org/Abstracts/are-datasets-for-nlme-models-large-enough-for-a-bootstrap-to-provide-reliable-parameter-uncertainty-distributions/}
}
\author{
Vipul Mann,  Matthew L. Fidler
}
