% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsingutil.R
\name{buildcovInfo}
\alias{buildcovInfo}
\title{Build covInfo list from varsVec and covarsVec}
\usage{
buildcovInfo(varsVec, covarsVec)
}
\arguments{
\item{varsVec}{character vector of variables that need to be added}

\item{covarsVec}{character vector of covariates that need to be added}
}
\value{
covInfo list of covariate info
}
\description{
Build covInfo list from varsVec and covarsVec
}
\author{
Vishal Sarsani
}
