% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsingutil.R
\name{buildupatedUI}
\alias{buildupatedUI}
\title{Build updated from the covariate and variable vector list}
\usage{
buildupatedUI(ui, varsVec, covarsVec, add = TRUE, indep = FALSE)
}
\arguments{
\item{ui}{compiled rxode2 nlmir2 model or fit}

\item{varsVec}{character vector of variables that need to be added}

\item{covarsVec}{character vector of covariates that need to be added}

\item{add}{boolean indicating if the covariate needs to be added or removed}

\item{indep}{a boolean indicating if the covariates should be added independently, or
sequentially (append to the previous model). only applicable to adding covariate}
}
\value{
updated ui with added covariates
}
\description{
Build updated from the covariate and variable vector list
}
\author{
Vishal Sarsani
}
