% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCM.R
\name{covarSearchAuto}
\alias{covarSearchAuto}
\title{Stepwise Covariate Model-selection (SCM) method}
\usage{
covarSearchAuto(...)
}
\value{
A list with elements \code{summaryTable} (combined forward and
backward results), \code{resFwd} (list of final fit and step table from
forward search), and \code{resBck} (same for backward search).
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{covarSearchAuto()} has been renamed to \code{\link[=runSCM]{runSCM()}}. This alias is provided
for backward compatibility and will be removed in a future release.
}
\keyword{internal}
