% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlmixrFormula.R
\name{.nlmixrFormulaDataPrep}
\alias{.nlmixrFormulaDataPrep}
\title{Perform any required data modifications for the nlmixrFormula interface}
\usage{
.nlmixrFormulaDataPrep(data, dvName, idName)
}
\arguments{
\item{data}{nlmixr data}

\item{dvName, idName}{The name of the DV and ID columns for the dataset,
respectively}
}
\value{
A data frame modified, as needed for nlmixrFormula; if data is NULL,
return NULL
}
\description{
Perform any required data modifications for the nlmixrFormula interface
}
