% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlmixrFormula.R
\name{.nlmixrFormulaParser}
\alias{.nlmixrFormulaParser}
\title{Parse the formula to extract the dependent variable, predictor, and random
effects}
\usage{
.nlmixrFormulaParser(object)
}
\arguments{
\item{object}{The formula to parse}
}
\value{
A list with names of "DV", "predictor", and "ranef" which are each
part of the formula broken down into calls
}
\description{
Parse the formula to extract the dependent variable, predictor, and random
effects
}
\examples{
.nlmixrFormulaParser(a~b+c~(c|id))
}
\keyword{Internal}
