% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlmixrFormula.R
\name{.nlmixrFormulaParserRanef}
\alias{.nlmixrFormulaParserRanef}
\title{Parse the random effects part of a formula}
\usage{
.nlmixrFormulaParserRanef(object)
}
\arguments{
\item{object}{The formula to parse}
}
\value{
An unnamed list with one element per random effect.  The elements
each have names of "ranefVar", "ranefGroup", and "start" indicating the
variable with the random effect, the grouping variable for the random
effect, and the starting value for the standard deviation of the random
effect.
}
\description{
Parse the random effects part of a formula
}
\examples{
\dontrun{
.nlmixrFormulaParserRanef(str2lang("(c|id)+(d|id2)"))
}
}
\keyword{Internal}
