% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlmixrFormula.R
\name{.nlmixrFormulaSetupIniFixed}
\alias{.nlmixrFormulaSetupIniFixed}
\title{Setup the ini() part of the model for fixed effects}
\usage{
.nlmixrFormulaSetupIniFixed(start, base = str2lang("{}"))
}
\arguments{
\item{start}{The starting estimates for the model}

\item{base}{The initial basis for the ini definition}
}
\value{
The inside of the ini() part of the model
}
\description{
Setup the ini() part of the model for fixed effects
}
\keyword{Internal}
