% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlmixrFormula.R
\name{.nlmixrFormulaSetupModel}
\alias{.nlmixrFormulaSetupModel}
\title{Setup the model() part of the model}
\usage{
.nlmixrFormulaSetupModel(
  start,
  predictor,
  residualModel,
  predictorVar = "value",
  data
)
}
\arguments{
\item{start}{The starting estimates (used when fixed effects need more
defition like for factors)}

\item{predictor}{The predictor from the formula}

\item{residualModel}{The residual model definition}

\item{predictorVar}{The variable in the data for the predictor}

\item{data}{The data used in the model}
}
\value{
the interior of the model()
}
\description{
Setup the model() part of the model
}
\keyword{Internal}
