% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_eq.R
\name{extract_eq.nlmixr2FitCore}
\alias{extract_eq.nlmixr2FitCore}
\alias{extract_eq.rxUi}
\title{Extract the equations from an nlmixr2/rxode2 model to produce a 'LaTeX'
equation.}
\usage{
\method{extract_eq}{nlmixr2FitCore}(
  model,
  intercept = "alpha",
  greek = "beta",
  greek_colors = NULL,
  subscript_colors = NULL,
  var_colors = NULL,
  var_subscript_colors = NULL,
  raw_tex = FALSE,
  swap_var_names = NULL,
  swap_subscript_names = NULL,
  ital_vars = FALSE,
  label = NULL,
  index_factors = FALSE,
  show_distribution = FALSE,
  wrap = FALSE,
  terms_per_line = 4,
  operator_location = "end",
  align_env = "aligned",
  use_coefs = FALSE,
  coef_digits = 2,
  fix_signs = TRUE,
  font_size = NULL,
  mean_separate,
  return_variances = FALSE,
  se_subscripts = FALSE,
  ...
)

\method{extract_eq}{rxUi}(
  model,
  intercept = "alpha",
  greek = "beta",
  greek_colors = NULL,
  subscript_colors = NULL,
  var_colors = NULL,
  var_subscript_colors = NULL,
  raw_tex = FALSE,
  swap_var_names = NULL,
  swap_subscript_names = NULL,
  ital_vars = FALSE,
  label = NULL,
  index_factors = FALSE,
  show_distribution = FALSE,
  wrap = FALSE,
  terms_per_line = 4,
  operator_location = "end",
  align_env = "aligned",
  use_coefs = FALSE,
  coef_digits = 2,
  fix_signs = TRUE,
  font_size = NULL,
  mean_separate,
  return_variances = FALSE,
  se_subscripts = FALSE,
  ...
)
}
\arguments{
\item{model}{The model to extract equations from}

\item{intercept, greek, greek_colors, subscript_colors, var_colors, var_subscript_colors, raw_tex, swap_var_names, swap_subscript_names, ital_vars, label, index_factors, show_distribution, wrap, terms_per_line, operator_location, align_env, use_coefs, coef_digits, fix_signs, font_size, mean_separate, return_variances, se_subscripts}{Ignored}
}
\description{
Extract the equations from an nlmixr2/rxode2 model to produce a 'LaTeX'
equation.
}
