% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit_printEquation.R
\name{extractEqHelper.if}
\alias{extractEqHelper.if}
\title{Generate LaTeX for if blocks}
\usage{
\method{extractEqHelper}{`if`}(x, ..., inModel, alignment, indent = 0L, firstIf = 0L, name)
}
\arguments{
\item{x}{language object to convert to LaTeX}

\item{...}{additional arguments passed to methods}

\item{inModel}{logical; whether the expression is inside a model block}

\item{indent}{tracks the current indention level}

\item{firstIf}{tracks if this is the first if at the current indent level}

\item{name}{optional name for the expression}
}
\description{
Generate LaTeX for if blocks
}
