% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCM.R
\name{forwardSearch}
\alias{forwardSearch}
\title{Forward covariate search}
\usage{
forwardSearch(
  varsVec,
  covarsVec,
  catvarsVec = NULL,
  fit,
  pVal = 0.05,
  outputDir,
  restart = FALSE
)
}
\arguments{
\item{varsVec}{character vector of variables that need to be added}

\item{covarsVec}{character vector of covariates that need to be added}

\item{catvarsVec}{character vector of categorical covariates that need to be added}

\item{fit}{an nlmixr2 'fit' object}

\item{pVal}{p-value that should be used for selecting covariates in the forward search}

\item{outputDir}{the name of the output directory that stores the covariate search result}

\item{restart}{a boolean that controls if the search should be restarted; default is FALSE}
}
\value{
returns the updated 'fit' object at the end of the forward search and a table of information for all the covariates tested
}
\description{
Forward covariate search
}
\author{
Vipul Mann, Matthew Fidler, Vishal Sarsani
}
