% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AICHelpers.R
\name{isBoundaryFit}
\alias{isBoundaryFit}
\title{Detect if a fit has any parameter at the boundary}
\usage{
isBoundaryFit(object)
}
\arguments{
\item{object}{A model fit to check}
}
\value{
\code{TRUE} if the fit is an nlmxir2 fit and it has a parameter at the
boundary; \code{FALSE} otherwise
}
\description{
Detect if a fit has any parameter at the boundary
}
