% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearizefocei.R
\name{isLinearizeMatch}
\alias{isLinearizeMatch}
\title{Check Linearization Match}
\usage{
isLinearizeMatch(fit, linFit, tol = 0.05)
}
\arguments{
\item{fit}{fit of nonlinear model.}

\item{linFit}{fit of linear model.}

\item{tol}{relative tolerance for matching. Default is 0.05.}
}
\value{
match list for OFV, omega, eta, and residual variance terms
}
\description{
Check Linearization Match
}
\author{
Omar Elashkar
}
