% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit_printEquation.R
\name{knit_print.nlmixr2FitCore}
\alias{knit_print.nlmixr2FitCore}
\alias{knit_print.rxUi}
\title{Extract the equations from an nlmixr2/rxode2 model to produce a 'LaTeX'
equation.}
\usage{
\method{knit_print}{nlmixr2FitCore}(x, ..., output = "equations")

\method{knit_print}{rxUi}(x, ...)
}
\arguments{
\item{x}{The model to extract equations from}

\item{...}{Ignored}

\item{output}{The type of output to request (currently, just "equations")}
}
\description{
Extract the equations from an nlmixr2/rxode2 model to produce a 'LaTeX'
equation.
}
