% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lassocov.R
\name{lassoCoefficients}
\alias{lassoCoefficients}
\title{Return Final lasso coefficients after finding optimal t}
\usage{
lassoCoefficients(
  fit,
  varsVec,
  covarsVec,
  catvarsVec,
  constraint = 1e-08,
  stratVar = NULL,
  ...
)
}
\arguments{
\item{fit}{nlmixr2 fit.}

\item{varsVec}{character vector of variables that need to be added}

\item{covarsVec}{character vector of covariates that need to be added}

\item{catvarsVec}{character vector of categorical covariates that need to be added}

\item{constraint}{theta cutoff. below cutoff then the theta will be fixed to zero}

\item{stratVar}{A variable to stratify on for cross-validation}

\item{...}{Other parameters to be passed to optimalTvaluelasso}
}
\value{
return data frame of final lasso coefficients
}
\description{
Return Final lasso coefficients after finding optimal t
}
\author{
Vishal Sarsani
}
