% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lasso.R
\name{lassoVar}
\alias{lassoVar}
\title{Title}
\usage{
lassoVar(
  fit,
  varsVec,
  covarsVec,
  covInformation = NULL,
  catCovariates = NULL,
  restart = FALSE
)
}
\arguments{
\item{fit}{}

\item{varsVec}{}

\item{covarsVec}{}

\item{covInformation}{}

\item{catCovariates}{}

\item{restart}{}
}
\description{
Title
}
