% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxUiLinearize.R
\name{linearizeErrorLines}
\alias{linearizeErrorLines}
\alias{linearizeErrorLines.norm}
\alias{linearizeErrorLines.default}
\alias{linearizeErrorLines.rxUi}
\title{This creates the linearization error lines from a rxui model}
\usage{
linearizeErrorLines(line)

\method{linearizeErrorLines}{norm}(line)

\method{linearizeErrorLines}{default}(line)

\method{linearizeErrorLines}{rxUi}(line)
}
\arguments{
\item{line}{line to parse}
}
\value{
error lines for posology
}
\description{
This creates the linearization error lines from a rxui model
}
\author{
Matthew L. Fidler
}
\keyword{internal}
