% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearizefocei.R
\name{linModGen}
\alias{linModGen}
\title{Generate a Linearization Model From Previous Fit}
\usage{
linModGen(ui, focei = TRUE, derivFct = FALSE)
}
\arguments{
\item{ui}{rx model or fit object.}

\item{focei}{boolean. If TRUE, use FOCEI linearization with individual and residual linearization. Default is TRUE.}

\item{derivFct}{boolean. If TRUE, use normalization derivative factors. Default is FALSE.}
}
\description{
Generate a Linearization Model From Previous Fit
}
\author{
Omar I. Elashkar
}
