% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AICHelpers.R
\name{listModelsTested}
\alias{listModelsTested}
\title{Make a list of models tested ready for reporting}
\usage{
listModelsTested(fitList, caption, excludeBoundary = TRUE, k = 2)
}
\arguments{
\item{fitList}{A named list of models that were tested}

\item{caption}{The caption for the resulting table}

\item{excludeBoundary}{Exclude a model from selection if it has a parameter
at its boundary}

\item{k}{numeric, the \emph{penalty} per parameter to be used; the
    default \code{k = 2} is the classical AIC.}
}
\value{
A data.frame with a "caption" attribute ready for printing in a
report with \code{pander::pander()}. The data.frame will have names of
"Description", "AIC", and "dAIC"; if exclusions are applied, the data.frame
will also have an "Exclude" column.
}
\description{
Make a list of models tested ready for reporting
}
