% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlmixr2extra-package.R
\docType{package}
\name{nlmixr2extra-package}
\alias{nlmixr2extra}
\alias{nlmixr2extra-package}
\title{nlmixr2extra: Nonlinear Mixed Effects Models in Population PK/PD, Extra Support Functions}
\description{
Fit and compare nonlinear mixed-effects models in differential equations with flexible dosing information commonly seen in pharmacokinetics and pharmacodynamics (Almquist, Leander, and Jirstrand 2015 \doi{10.1007/s10928-015-9409-1}). Differential equation solving is by compiled C code provided in the 'rxode2' package (Wang, Hallow, and James 2015 \doi{10.1002/psp4.12052}). This package is for support functions like preconditioned fits \doi{10.1208/s12248-016-9866-5}, boostrap and stepwise covariate selection.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://nlmixr2.github.io/nlmixr2extra/}
  \item \url{https://github.com/nlmixr2/nlmixr2extra/}
  \item Report bugs at \url{https://github.com/nlmixr2/nlmixr2extra/issues/}
}

}
\author{
\strong{Maintainer}: Matthew Fidler \email{matthew.fidler@gmail.com} (\href{https://orcid.org/0000-0001-8538-6691}{ORCID})

Authors:
\itemize{
  \item Vipul Mann \email{vm2583@columbia.edu}
  \item Vishal Sarsani \email{vsarsani@umass.edu} (\href{https://orcid.org/0000-0002-9272-3438}{ORCID})
  \item Bill Denney \email{wdenney@humanpredictions.com} (\href{https://orcid.org/0000-0002-5759-428X}{ORCID})
}

Other contributors:
\itemize{
  \item Christian Bartels \email{christian.bartels@novartis.com} [contributor]
  \item Omar Elashkar \email{omar.i.elashkar@gmail.com} (\href{https://orcid.org/0000-0002-5505-778X}{ORCID}) [contributor]
}

}
\keyword{internal}
