% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlmixr_formula.R
\name{nlmixr_formula}
\alias{nlmixr_formula}
\title{A simple formula-based interface for nlmixr2}
\usage{
nlmixr_formula(
  object,
  data,
  start,
  param = NULL,
  ...,
  residual_model = ~add(add_err)
)
}
\arguments{
\item{object}{The formula defining the model (see details)}

\item{data}{The data to fit}

\item{start}{A named list of starting estimates.  The names define the
parameters in the model.  If a single parameter estimate is desired, it can
be given here.  If a parameter estimate per factor level is desired, either
a single starting estimate can be given across all factor levels or one
estimate may be given per factor level.  (Specify the factors with the \code{param} argument.)}

\item{...}{Passed to \code{nlmixr()}}

\item{residual_model}{The residual model formula to use.}
}
\value{
The model fit from \code{nlmixr()}
}
\description{
A simple formula-based interface for nlmixr2
}
\details{
The formula is given with different notation than typical formulas.  The
formula notation is inspired by and similar to \code{lme4::nlmer()}.  It is a
3-part formula: \code{dependent_variable~predictor_equation~random_effects}.

The \code{dependent_variable} is any variable in the dataset.  It may not
include any math; for example, \code{log(DV)} is not allowed.

The \code{predictor_equation} is any valid math, and it will be used directly
in the nlmixr2 model.

The \code{random_effects} are one or more random effect parameters defined by
putting the parameter in parentheses and putting a vertical bar and the
grouping parameter.  Only one grouping parameter is allowed for all random
effects.  An example would be \code{(slope|ID)} to estimate a random effect
parameter named "slope" for each "ID" in the data.
}
\examples{
nlmixr_formula(
  height ~ (Asym+Asym_re)+(R0-(Asym+Asym_re))*exp(-exp(lrc)*age) ~ (Asym_re|Seed),
  data = Loblolly,
  start = list(Asym = 103, R0 = -8.5, lrc = -3.3, add_err=1),
  est="focei"
)
}
