% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computingutil.R
\name{optimUnisampling}
\alias{optimUnisampling}
\title{Sample from uniform distribution by optim}
\usage{
optimUnisampling(xvec, N = 1000, medValue, floorT = TRUE)
}
\arguments{
\item{xvec}{A vector of min,max values . Ex:c(10,20)}

\item{N}{Desired number of values}

\item{medValue}{Desired Median}

\item{floorT}{boolean indicating whether to round up}
}
\value{
Samples with approx desired median.
}
\description{
Sample from uniform distribution by optim
}
\examples{
# Simulate 1000 creatine clearance values with median of 71.7 within range of c(6.7,140)
creatCl <- optimUnisampling(xvec=c(6.7,140), N=1000, medValue = 71.7, floorT=FALSE)
}
\author{
Vishal Sarsani
}
