% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multistart.R
\name{plot.nlmixr2Multistart}
\alias{plot.nlmixr2Multistart}
\title{Plot a multistart result}
\usage{
\method{plot}{nlmixr2Multistart}(x, type = c("waterfall", "parameters"), kBest = 20L, dOfvMax = NULL, ...)
}
\arguments{
\item{x}{A \code{nlmixr2Multistart} object from \code{\link[=multistart]{multistart()}}}

\item{type}{\code{"waterfall"} plots each start's objective function relative to
the best one, worst to best; \code{"parameters"} plots how each parameter
estimate varies across the best starts.}

\item{kBest}{Number of starts to show in the \code{"parameters"} plot}

\item{dOfvMax}{Upper limit for the waterfall's objective-function axis, for
zooming in when one start is far worse than the rest}

\item{...}{ignored}
}
\value{
A ggplot2 object
}
\description{
Plot a multistart result
}
\seealso{
Other Multistart:
\code{\link[=multistart]{multistart()}},
\code{\link[=multistartControl]{multistartControl()}}
}
\concept{Multistart}
