% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{profileNlmixr2FitCoreRet}
\alias{profileNlmixr2FitCoreRet}
\title{Give the output data.frame for a single model for profile.nlmixr2FitCore}
\usage{
profileNlmixr2FitCoreRet(fitted, which, fixedVal, rowname = 0)
}
\arguments{
\item{fitted}{The fit model}

\item{which}{The parameter names to perform likelihood profiling on
(\code{NULL} indicates all parameters)}
}
\value{
A data.frame with columns named "Parameter" (the parameter name(s)
that were fixed), OFV (the objective function value), and the current
estimate for each of the parameters.
}
\description{
Give the output data.frame for a single model for profile.nlmixr2FitCore
}
\seealso{
Other Profiling: 
\code{\link{fixedControl}()},
\code{\link{llpControl}()},
\code{\link{profile.nlmixr2FitCore}()},
\code{\link{profileFixed}()},
\code{\link{profileLlp}()}
}
\concept{Profiling}
