% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{profileNlmixr2FitDataNewEst}
\alias{profileNlmixr2FitDataNewEst}
\title{Generate a new estimate for likelihood profiling, and return a list
indicating if the new estimate should not be run (because the current data
find the value within the tolerance)}
\usage{
profileNlmixr2FitDataNewEst(
  estimates,
  which,
  direction,
  bound,
  ofvIncrease,
  method = "linapprox"
)
}
\arguments{
\item{estimates}{A data.frame with the parameter estimates}
}
\description{
Generate a new estimate for likelihood profiling, and return a list
indicating if the new estimate should not be run (because the current data
find the value within the tolerance)
}
