% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{profileNlmixr2SingleParam}
\alias{profileNlmixr2SingleParam}
\alias{profileNlmixr2MultiParam}
\title{Estimate the objective function value for a model while fixing a single set
of defined parameter values (for use in parameter profiling)}
\usage{
profileNlmixr2SingleParam(fitted, which, returnType = c("OFVdf", "model"))

profileNlmixr2MultiParam(fitted, which)
}
\arguments{
\item{fitted}{The fit model}

\item{which}{A data.frame with column names of parameters to fix and values
of the fitted value to fix (one row per set of parameters to estimate)}
}
\value{
\code{which} with a column named \code{OFV} added with the objective function
value of the fitted estimate fixing the parameters in the other columns
}
\description{
Estimate the objective function value for a model while fixing a single set
of defined parameter values (for use in parameter profiling)
}
\section{Functions}{
\itemize{
\item \code{profileNlmixr2MultiParam()}: Estimate the objective function values
for a model while fixing defined parameter values (for use in parameter
profiling)

}}
\seealso{
Other Profiling: 
\code{\link{profile.nlmixr2FitCore}()}
}
\concept{Profiling}
