% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit_printEquation.R, R/reexport.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{knit_print}
\alias{ini}
\alias{model}
\alias{nlmixrWithTiming}
\alias{nlmixr}
\alias{nlmixr2}
\alias{setCov}
\alias{rxUiDeparse}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{knitr}{\code{\link[knitr:knit_print]{knit_print()}}}

  \item{nlmixr2est}{\code{\link[nlmixr2est:nlmixr]{nlmixr()}}, \code{\link[nlmixr2est:nlmixr2]{nlmixr2()}}, \code{\link[nlmixr2est:nlmixrWithTiming]{nlmixrWithTiming()}}, \code{\link[nlmixr2est:setCov]{setCov()}}}

  \item{rxode2}{\code{\link[rxode2:ini]{ini()}}, \code{\link[rxode2:model]{model()}}, \code{\link[rxode2:rxUiDeparse]{rxUiDeparse()}}}
}}

