% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lassocov.R
\name{regularmodel}
\alias{regularmodel}
\title{Regular lasso model}
\usage{
regularmodel(
  fit,
  varsVec,
  covarsVec,
  catvarsVec,
  constraint = 1e-08,
  lassotype = "regular",
  stratVar = NULL,
  ...
)
}
\arguments{
\item{fit}{nlmixr2 fit.}

\item{varsVec}{character vector of variables that need to be added}

\item{covarsVec}{character vector of covariates that need to be added}

\item{catvarsVec}{character vector of categorical covariates that need to be added}

\item{constraint}{theta cutoff. below cutoff then the theta will be fixed to zero.}

\item{lassotype}{must be  'regular' , 'adaptive', 'adjusted'}

\item{stratVar}{A variable to stratify on for cross-validation.}

\item{...}{Other parameters to be passed to optimalTvaluelasso}
}
\value{
return fit of the regular coefficients
}
\description{
Regular lasso model
}
\author{
Vishal Sarsani
}
