% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iivSearch.R
\name{rerunTopN.linIIVSearch}
\alias{rerunTopN.linIIVSearch}
\title{Rerun Top N Original Models From IIV Search}
\usage{
\method{rerunTopN}{linIIVSearch}(x, n = 5)
}
\arguments{
\item{x}{Linearized iivSearch results}

\item{n}{number of models to rerun}
}
\description{
Rerun Top N Original Models From IIV Search
}
\author{
Omar I. Elashkar
}
