% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resSearch.R
\name{resSearch}
\alias{resSearch}
\title{Exhaustively Search for Residual Error Model}
\usage{
resSearch(fit)
}
\arguments{
\item{fit}{Fitted model object}
}
\value{
list of original model and summary for different residual models searched
}
\description{
Exhaustively Search for Residual Error Model
}
\author{
Omar I. Elashkar
}
