% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlmixrFormula.R
\name{.nlmixrFormulaExpandStartParam}
\alias{.nlmixrFormulaExpandStartParam}
\alias{.nlmixrFormulaExpandStartParamSingle}
\title{Expand parameters to include their factor representations, if applicable.}
\usage{
.nlmixrFormulaExpandStartParam(start, param, data)

.nlmixrFormulaExpandStartParamSingle(startName, startValue, param, data)
}
\arguments{
\item{start}{the starting values for the model}

\item{param}{The parameter in the model}

\item{data}{The dataset}

\item{startName}{The base name for the parameter}

\item{startValue}{The initial value for the base parameter}
}
\value{
A list with the ini and model parts needed to use the parameter
}
\description{
Expand parameters to include their factor representations, if applicable.
}
\section{Functions}{
\itemize{
\item \code{.nlmixrFormulaExpandStartParamSingle()}: Expand a single parameter in a model using dataset factors, if applicable

}}
\keyword{Internal}
