% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlmixrFormula.R
\name{.renameOrOverwrite}
\alias{.renameOrOverwrite}
\title{Rename a column in a dataset, optionally overwriting it if the column does
not exist}
\usage{
.renameOrOverwrite(data, newName, oldName)
}
\arguments{
\item{data}{The dataset to modify}

\item{newName, oldName}{The new and old column names}
}
\value{
data with \code{data[[newName]] <- data[[charOld]]}
}
\description{
Rename a column in a dataset, optionally overwriting it if the column does
not exist
}
\examples{
\dontrun{
.renameOrOverwrite(data.frame(A=1), newName="B", oldName="A")
}
}
\keyword{Internal}
