% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{fixedControl}
\alias{fixedControl}
\title{Control options for fixed-value likelihood profiling}
\usage{
fixedControl()
}
\value{
A validated list of control options for fixed-value likelihood
profiling
}
\description{
Control options for fixed-value likelihood profiling
}
\seealso{
\code{\link[=profileFixed]{profileFixed()}}

Other Profiling: 
\code{\link{llpControl}()},
\code{\link{profile.nlmixr2FitCore}()},
\code{\link{profileFixed}()},
\code{\link{profileLlp}()},
\code{\link{profileNlmixr2FitCoreRet}()}
}
\concept{Profiling}
