% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesiancovsel.R
\name{horseshoeSummardf}
\alias{horseshoeSummardf}
\title{Create Horseshoe summary posterior estimates}
\usage{
horseshoeSummardf(fit, covarsVec, ...)
}
\arguments{
\item{fit}{compiled rxode2 nlmir2 model fit}

\item{covarsVec}{character vector of covariates that need to be added}

\item{...}{other parameters passed to brm(): warmup = 1000, iter = 2000, chains = 4, cores = 4,
control = list(adapt_delta = 0.99, max_treedepth = 15)}
}
\value{
Horse shoe Summary data frame of all covariates
}
\description{
Create Horseshoe summary posterior estimates
}
\author{
Vishal Sarsani, Christian Bartels
}
