% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iivSearch.R
\name{iivSearch}
\alias{iivSearch}
\title{Automated Inter-Individual Variability Search}
\usage{
iivSearch(fit, ...)
}
\arguments{
\item{fit}{a model fit}
}
\description{
Automated Inter-Individual Variability Search
}
\author{
Omar I. Elashkar
}
