% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearizefocei.R
\name{linearize}
\alias{linearize}
\title{Perform linearization of a model fitted using FOCEI}
\usage{
linearize(
  fit,
  mceta = c(-1, 10, 100, 1000),
  relTol = 0.4,
  focei = NA,
  derivFct = FALSE,
  plot = FALSE
)
}
\arguments{
\item{fit}{fit of nonlinear model.}

\item{mceta}{a numeric vector for mceta to try. See decription.}

\item{relTol}{relative deviation tolerance between original and linearized models objective functions. Used for switching if focei = NA. See details.}

\item{focei}{Default is NA for automatic switch from FOCEI to FOCE if failed. See details.}

\item{derivFct}{boolean. If TRUE, turn on derivatives for linearization. Default is FALSE.}

\item{plot}{boolean. Print plot of linearized vs original}
}
\description{
Perform linearization of a model fitted using FOCEI
}
\details{
mceta vector will be iterated over to find the best linearization if linearization failed.
Escalating to next mceta will depend on the relative deviation \code{relTol} of the original and linearized models objective functions.

If \code{focei} is set to \code{NA}, the function will first try to linearize using FOCEI.
If the relative deviation between original and linearized models objective functions is greater than \code{relTol}, it will switch to FOCE where residual linearization is skipped.
If \code{focei} is set to \code{TRUE}, the function will use FOCEI linearization with individual and residual linearization.
If \code{focei} is set to \code{FALSE}, the function will use FOCE linearization with residual linearization skipped.

If \code{derivFct} is set to \code{TRUE}, the function will use derivatives for linearization.
This might be usefull to try with FOCEI.

\code{plot} argument can only print ggplot figure with default settings.
If a user wish to capture the plot, one might use \code{linearizePlot()} call.
}
\author{
Omar Elashkar
}
