% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearizefocei.R
\name{linearize}
\alias{linearize}
\title{Perform linearization of a model fitted using FOCEI}
\usage{
linearize(
  fit,
  mceta = c(-1, 10, 100, 1000),
  relTol = 0.25,
  focei = NA,
  addEtas = FALSE,
  derivFct = FALSE,
  plot = FALSE,
  est = "focei"
)
}
\arguments{
\item{fit}{fit of nonlinear model fitted using any method with at least one eta. See details.}

\item{mceta}{a numeric vector for mceta to try. See details.}

\item{relTol}{relative deviation tolerance between original and linearized models objective functions. Used for switching if focei = NA. See details.}

\item{focei}{Default is NA for automatic switch from FOCEI to FOCE if failed. See details.}

\item{addEtas}{boolean. If TRUE, add etas on every theta and fix it to small value to get derivatives. Default is FALSE.}

\item{derivFct}{boolean. If TRUE, turn on derivatives for linearization. Default is FALSE.}

\item{plot}{boolean. Print plot of linearized vs original}

\item{est}{Character. The estimation method used for the linearized model fit. Only 'focei' is supported.}
}
\value{
a fit object with subclass nlmixr2linearize
}
\description{
Perform linearization of a model fitted using FOCEI
}
\details{
The function accepts a fit object fitted using any method. However, if the method is not FOCE+I, the model is going to be evaluated first.

\code{mceta} vector will be iterated over to find the best linearization if linearization failed.
Escalating to next mceta will depend on the relative deviation \code{relTol} of the original and linearized models objective functions.

If \code{focei} is set to \code{NA}, the function will first try to linearize using FOCEI.
If the relative deviation between original and linearized models objective functions is greater than \code{relTol}, it will switch to FOCE where residual linearization is skipped.
If \code{focei} is set to \code{TRUE}, the function will use FOCEI linearization with individual and residual linearization.
If \code{focei} is set to \code{FALSE}, the function will use FOCE linearization with residual interaction linearization skipped.

If \code{derivFct} is set to \code{TRUE}, the function will use an extra factor for linearization that might help in stabilization.
This might be useful to try with FOCEI.

\code{plot} argument can only print ggplot figure with default settings.
If a user wish to capture the plot, one might use \code{linearizePlot()} call.
}
\author{
Omar I. Elashkar
}
