% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearizefocei.R
\name{linearizePlot}
\alias{linearizePlot}
\title{Plot Original Versus Linear Models iObj and Etas}
\usage{
linearizePlot(nl, lin)
}
\arguments{
\item{nl}{non-linear fitting object}

\item{l}{linear fitting object}
}
\value{
ggplot object
}
\description{
Plot Original Versus Linear Models iObj and Etas
}
\author{
Omar Elashkar
}
