% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearizefocei.R
\name{linearizePlot}
\alias{linearizePlot}
\title{Plot Original Versus Linear Models iObj and Etas}
\usage{
linearizePlot(lin)
}
\arguments{
\item{l}{linear fitting object}
}
\value{
ggplot object
}
\description{
Plot Original Versus Linear Models iObj and Etas
}
\details{
If the non-linear fit was not FOCEI, the objective functions are not comparable, hence can be ignored upon comparison.
}
\author{
Omar I. Elashkar
}
