% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{profile.nlmixr2FitCore}
\alias{profile.nlmixr2FitCore}
\title{Perform likelihood profiling on nlmixr2 focei fits}
\usage{
\method{profile}{nlmixr2FitCore}(fitted, ..., which = NULL, method = "llp", control = list())
}
\arguments{
\item{fitted}{The fit model}

\item{...}{ignored}

\item{which}{The parameter names to perform likelihood profiling on
(\code{NULL} indicates all parameters)}

\item{method}{Method to use for profiling}

\item{control}{Control arguments for the \code{method}}
}
\value{
A data.frame with one column named \code{Parameter} indicating the
parameter being fixed on that row, one column for the \code{OFV} indicating the
OFV estimated for the model at that step, one column named \code{profileBound}
indicating the estimated value for the profile likelihood and its step
above the minimum profile likelihood value, and columns for each parameter
estimate (or fixed) in the model.
}
\description{
The search will stop when either the OFV is within \code{ofvtol} of the desired
OFV change or when the parameter is interpolating to more significant digits
than specified in \code{paramDigits}.
}
\seealso{
Other Profiling: 
\code{\link{profileNlmixr2SingleParam}()}
}
\concept{Profiling}
