% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{profile.nlmixr2FitCore}
\alias{profile.nlmixr2FitCore}
\title{Perform likelihood profiling on nlmixr2 focei fits}
\usage{
\method{profile}{nlmixr2FitCore}(
  fitted,
  ...,
  which = NULL,
  maxpts = 10,
  ofvIncrease = qchisq(0.95, df = 1),
  normQuantile = qnorm(p = 0.975),
  rse_theta,
  tol = 0.001,
  ignoreBounds = FALSE,
  quiet = TRUE
)
}
\arguments{
\item{fitted}{The fit model}

\item{...}{ignored}

\item{which}{The parameter names to perform likelihood profiling on
(\code{NULL} indicates all parameters)}

\item{maxpts}{The maximum number of fits to perform in each direction (higher
or lower values) for each parameter.  The total maximum number of estimates
that may be performed is \code{maxpts*2*length(which)}.}

\item{ofvIncrease}{The targetted change in objective function value (3.84
corresponds to a Chi-squared test with a 95\% confidence interval)}

\item{normQuantile}{The quantile of a normal distribution to use for the
initial estimates.}

\item{rse_theta}{The relative standard error (percent) for the model
parameters.  It can be missing (the default) in which case a default value
of 30\% will be applied.  If given as a single number, it will be applied to
all parameters.  If given as a named vector of numbers, it will be applied
to each named parameter.}

\item{tol}{The relative tolerance for the parameter estimate being close
enough to the \code{ofvIncrease}.  The tolerance is applied to the
parameter estimate and not the the \code{ofvIncrease} to allow for ranges
with a high derivative to complete in a reasonable number of iterations.}

\item{ignoreBounds}{Should the boundary conditions in the model be ignored?}

\item{quiet}{Print less information from model estimation steps}
}
\value{
A data.frame with one column for the \code{Parameter} name being
fixed on that row, one column for the \code{OFV}, and columns for each
parameter estimate (or fixed value) in the model.
}
\description{
Perform likelihood profiling on nlmixr2 focei fits
}
