% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{profile.nlmixr2FitCore}
\alias{profile.nlmixr2FitCore}
\title{Perform likelihood profiling on nlmixr2 focei fits}
\usage{
\method{profile}{nlmixr2FitCore}(
  fitted,
  ...,
  which = NULL,
  maxpts = 10,
  ofvIncrease = qchisq(0.95, df = 1),
  normQuantile = qnorm(p = 0.975),
  rseTheta,
  itermax = 10,
  ofvtol = 0.005,
  paramDigits = 3
)
}
\arguments{
\item{fitted}{The fit model}

\item{...}{ignored}

\item{which}{The parameter names to perform likelihood profiling on
(\code{NULL} indicates all parameters)}

\item{maxpts}{The maximum number of fits to perform in each direction (higher
or lower values) for each parameter.  The total maximum number of estimates
that may be performed is \code{maxpts*2*length(which)}.}

\item{ofvIncrease}{The targetted change in objective function value (3.84
corresponds to a Chi-squared test with a 95\% confidence interval)}

\item{normQuantile}{The quantile of a normal distribution to use for the
initial estimates.}

\item{rseTheta}{The relative standard error (percent) for the model
parameters.  It can be missing (the default) in which case a default value
of 30\% will be applied.  If given as a single number, it will be applied to
all parameters.  If given as a named vector of numbers, it will be applied
to each named parameter.}

\item{ofvtol}{The relative tolerance for the objective function being close
enough to the \code{ofvIncrease}.}

\item{paramDigits}{The number of significant digits required for the
parameter.  When interpolation attempts to get smaller than that number of
significant digits, it will stop.}
}
\value{
A data.frame with one column named \code{Parameter} indicating the
parameter being fixed on that row, one column for the \code{OFV} indicating the
OFV estimated for the model at that step, one column named \code{profileBound}
indicating the estimated value for the profile likelihood and its step
above the minimum profile likelihood value, and columns for each parameter
estimate (or fixed) in the model.
}
\description{
The search will stop when either the OFV is within \code{ofvtol} of the desired
OFV change or when the parameter is interpolating to more significant digits
than specified in \code{paramDigits}.
}
\seealso{
Other Profiling: 
\code{\link{profileNlmixr2SingleParam}()}
}
\concept{Profiling}
