% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_eq.R, R/reexport.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{extract_eq}
\alias{ini}
\alias{model}
\alias{nlmixrWithTiming}
\alias{nlmixr}
\alias{nlmixr2}
\alias{setCov}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{equatiomatic}{\code{\link[equatiomatic]{extract_eq}}}

  \item{nlmixr2est}{\code{\link[nlmixr2est:nlmixr2]{nlmixr}}, \code{\link[nlmixr2est]{nlmixr2}}, \code{\link[nlmixr2est]{nlmixrWithTiming}}, \code{\link[nlmixr2est]{setCov}}}

  \item{rxode2}{\code{\link[rxode2]{ini}}, \code{\link[rxode2]{model}}}
}}

