% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iivSearch.R
\name{rerunTopN}
\alias{rerunTopN}
\title{Rerun Top N Original Models From A Search}
\usage{
rerunTopN(x, ...)
}
\arguments{
\item{x}{a Search object}

\item{...}{Other Parameters}
}
\description{
Rerun Top N Original Models From A Search
}
\author{
Omar I. Elashkar
}
